LLaVA-NeXT/scripts/inf/requirements_inf.txt
nltk==3.8.1
rouge-score==0.1.2
bert-score==0.3.13
torch>=2.0.0
numpy>=1.20.0
tqdm>=4.64.0
transformers>=4.30.0
Pillow>=9.0.0